Tweak the ftplugin
Turn off modelines for snippets files. We don't need them, and it causes
trouble if the snippets file is for the vim filetype and modelines are
present in individual snippets.

Update the ftplugin to behave better. This means use b:did_ftplugin,
toggle cpo as appropriate, and define a b:undo_ftplugin.